+++ src/main/java/at/hannibal2/skyhanni/features/event/diana/InquisitorWaypointShare.kt
@@ -20,7 +20,7 @@ import org.lwjgl.input.Keyboard
 import kotlin.time.Duration.Companion.seconds
 
 object InquisitorWaypointShare {
-    private val config get() = SkyHanniMod.feature.diana.inquisitorSharing
+    private val config get() = SkyHanniMod.feature.event.diana.inquisitorSharing
     private val partyPattern =
         "§9Party §8> (?<playerName>.*)§f: §rx: (?<x>-?[0-9]{1,4}), y: (?<y>-?[0-9]{1,4}), z: (?<z>-?[0-9]{1,4})\\b".toPattern()
     private val diedPattern = "§9Party §8> (?<playerName>.*)§f: §rInquisitor dead!".toPattern()